Ensuring (int) overflow doesn't mess up allocation of, or indexing into, distance matrices.  Replacing int with size_t in: Variables in checkZeroDist, computeMLDist (in phyloanalysis.cpp) Variables in Alignment::printDist (pos). Variables in PhyloTree::correctDist (nsqr, pos, i, k). Variables in PhyloTree::computeDist (num_pairs, pos, seq1). Variables in 4-parameter PhyloTree::computeDist (n, nSquared, i). Variables in PhyloTree::computeObsDist (pos, seq1, seq2). Parameters to Alignment member functions: computeStateFreq, getSeqName, isGapOnlySeq Note, Alignment::checkIndenticalSeq should use hashing. Values returned by Alignment member functions: getNSeq Other places there might have been issues...PhyloTree::initializeAllPartialPars
While I was there I made some other Alignment member functions return size_t
(e.g. getNPattern), changed the parameter to isGapOnlySeq to size_t.
Also: denominator calculated in PhyloTree::computeDist_Experimental
now matches that used in the non-experimental distance computation
(so the JC distances they report are consistent, to about four decimal places).

+++ alignment/alignment.h
@@ -335,21 +335,21 @@ public:
     /**
             @return number of sequences
      */
-    inline int getNSeq() const {
+    inline size_t getNSeq() const {
         return seq_names.size();
     }
 
     /**
             @return number of sites (alignment columns)
      */
-    inline int getNSite() {
+    inline size_t getNSite() {
         return site_pattern.size();
     }
 
     /**
              @return number of patterns
      */
-    inline int getNPattern() {
+    inline size_t getNPattern() {
         return size();
     }
 
@@ -443,7 +443,7 @@ public:
             @return TRUE if seq_id contains only gaps or missing characters
             @param seq_id sequence ID
      */
-    bool isGapOnlySeq(int seq_id);
+    bool isGapOnlySeq(size_t seq_id);
 
     virtual bool isSuperAlignment() {
         return false;
